Initial unit tests
