+++ src/VSIXBootstrapper.Shared/CoInitializer.h
@@ -5,12 +5,26 @@
 
 #pragma once
 
+struct CoInitializerTraits
+{
+    static HRESULT __cdecl Initialize(_In_ LPVOID)
+    {
+        return ::CoInitialize(NULL);
+    };
+
+    static void __cdecl Uninitialize()
+    {
+        return ::CoUninitialize();
+    };
+};
+
+template <class _Traits = CoInitializerTraits>
 class CoInitializer
 {
 public:
     CoInitializer()
     {
-        auto hr = ::CoInitialize(NULL);
+        auto hr = _Traits::Initialize(NULL);
         if (FAILED(hr))
         {
             throw win32_error(hr);
@@ -19,7 +33,7 @@ public:
 
     ~CoInitializer()
     {
-        ::CoUninitialize();
+        _Traits::Uninitialize();
     }
 
 private:
