+++ src/VSIXBootstrapper.Shared/Exceptions.h
