+++ src/VSIXBootstrapper.Shared/Process.cpp
