+++ src/VSIXBootstrapper.Shared/Process.h
