+++ src/VSIXBootstrapper.Shared/RegistryKey.cpp
