+++ src/VSIXBootstrapper.Shared/RegistryKey.h
