+++ src/VSIXBootstrapper.Shared/Resources.cpp
