+++ src/VSIXBootstrapper.Shared/Resources.h
