+++ src/VSIXBootstrapper.Shared/VSIXBootstrapper.h
@@ -1,0 +1,8 @@
+// <copyright file="CoInitializer.h" company="Microsoft Corporation">
+// Copyright (C) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt in the project root for license information.
+// </copyright>
+
+#pragma once
+
+int Run(_In_ HINSTANCE hInstance, _In_ LPWSTR lpCmdLine, _In_ int nCmdShow);
