+++ test/VSIXBootstrapper.Test/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="MicroBuild.Core" version="0.2.0" targetFramework="native" developmentDependency="true" />
+  <package id="Microsoft.VisualStudio.Setup.Configuration.Native" version="1.5.125-rc" targetFramework="native" developmentDependency="true" />
+  <package id="Nerdbank.GitVersioning" version="1.5.62" targetFramework="native" developmentDependency="true" />
+</packages>
