+++ test/VSIXBootstrapper.Test/stdafx.cpp
@@ -1,0 +1,6 @@
+// <copyright file="stdafx.cpp" company="Microsoft Corporation">
+// Copyright (C) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt in the project root for license information.
+// </copyright>
+
+#include "stdafx.h"
