+++ test/VSIXBootstrapper.Test/stdafx.h
@@ -1,0 +1,16 @@
+// <copyright file="stdafx.h" company="Microsoft Corporation">
+// Copyright (C) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE.txt in the project root for license information.
+// </copyright>
+
+#pragma once
+
+#include "targetver.h"
+
+// Reference project header files
+#include <stdafx.h>
+
+// C++ unit test header files
+#include "CppUnitTest.h"
+
+// Project header files
